Updated code to run on a per-year or proj_group basis
As a result, there's a tonne more graphs generated.

Also messed with excel a bit more.

Still need to add in grouped bar charts. I thikn instead of trying to do everyhting in one function it will be cleaner to just take the ax from the subplot function and directly use matplotlib.

Co-Authored-By: Rowena-ycj <rowena-ycj@users.noreply.github.com>